VAS products management update

+++ app/Services/VasProductService.php
@@ -6,8 +6,10 @@ use App\Repositories\VasProductRepository;
 use App\Traits\CrudTrait;
 use App\Traits\FileTrait;
 use App\Traits\RedisTrait;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
+use Illuminate\Support\Facades\Redis;
 
 
 class VasProductService
@@ -26,17 +28,24 @@ class VasProductService
 
     public function getVasProducts()
     {
-        return $this->vasProductRepository->findAll(null,null, ['column'=> 'created_at', 'direction' => 'desc']);
+        return $this->vasProductRepository->findAll(null, null, ['column'=> 'display_order', 'direction' => 'asc']);
     }
 
     public function storeVasProducts($request)
     {
         try {
             DB::transaction(function () use ($request) {
+                $vas_products = $this->vasProductRepository->vasProducts();
+                $i = 1;
+                if (!empty($vas_products)) {
+                    $i = $vas_products->display_order + 1;
+                }
 
                 if (request()->hasFile('image')) {
                     $request['image'] = 'storage/' . $request['image']->storeAs('vas-product', time() . '-' . bin2hex(random_bytes(4)) . '-' . $request['image']->getClientOriginalName());
                 }
+                $request['display_order'] = $i;
+
                 $this->save($request);
 
             });
@@ -56,7 +65,7 @@ class VasProductService
             DB::transaction(function () use ($data, $id, $vasProduct) {
                 if (request()->hasFile('image')) {
                     $data['image'] = 'storage/' . $data['image']->storeAs('vas-product', time() . '-' . bin2hex(random_bytes(4)) . '-' . $data['image']->getClientOriginalName());
-                    $this->deleteFile($vasProduct->image);
+                    if($vasProduct->image) $this->deleteFile($vasProduct->image);
                 }
                 $vasProduct->update($data);
             });
@@ -80,5 +89,12 @@ class VasProductService
 
         //$this->redisDel(self::VAS_PRODUCT_REDIS_KEY);
         return Response('VAS Product has been successfully deleted');
+    }
+
+    public function tableSortable($data)
+    {
+        $this->vasProductRepository->vasProductsTableSort($data);
+        //$this->redisDel(self::VAS_PRODUCT_REDIS_KEY);
+        return new Response('Sequence has been successfully update');
     }
 }
